+++ tests/test_generate_sw_disp_tile.py
@@ -1,8 +1,10 @@
 from collections import namedtuple
 from datetime import datetime
 from unittest import mock
+from unittest.mock import patch
 
 import rioxarray  # noqa
+import xarray as xr
 from osgeo import gdal
 
 from opera_disp_tms import generate_sw_disp_tile as sw
@@ -72,3 +74,34 @@ def test_find_needed_granules():
         needed_granules = sw.find_needed_granules([1], datetime(2021, 1, 1), datetime(2021, 1, 3), strategy='all')
         assert len(needed_granules[1]) == 3
         assert needed_granules[1] == granules
+
+
+def test_update_reference_date():
+    def make_xr(value, ref_date):
+        dim_coords = dict(dims=['x'], coords={'x': [1]})
+        da = {'short_wavelength_displacement': xr.DataArray([value], **dim_coords)}
+        out = xr.Dataset(da, attrs={'reference_date': ref_date, 'bbox': [0, 0, 0, 0]})
+        return out
+
+    to_correct = make_xr(2, datetime(2021, 1, 1))
+    back1 = make_xr(4, datetime(2020, 1, 1))
+    back2 = make_xr(7, datetime(2019, 1, 1))
+
+    frame = sw.FrameMeta(1, datetime(2020, 1, 1), (0, 0))
+    pkg = 'opera_disp_tms.generate_sw_disp_tile'
+    with patch(f'{pkg}.find_needed_granules') as mock_find, patch(f'{pkg}.load_sw_disp_granule') as mock_load:
+        mock_find.return_value = {1: [None]}
+        mock_load.side_effect = [back1, back2]
+        to_correct = sw.update_reference_date(to_correct, frame)
+        assert to_correct.attrs['reference_date'] == datetime(2020, 1, 1)
+        assert to_correct.short_wavelength_displacement.values == 6
+
+    to_correct = make_xr(2, datetime(2021, 1, 1))
+    frame = sw.FrameMeta(1, datetime(2019, 1, 1), (0, 0))
+    pkg = 'opera_disp_tms.generate_sw_disp_tile'
+    with patch(f'{pkg}.find_needed_granules') as mock_find, patch(f'{pkg}.load_sw_disp_granule') as mock_load:
+        mock_find.return_value = {1: [None]}
+        mock_load.side_effect = [back1, back2]
+        to_correct = sw.update_reference_date(to_correct, frame)
+        assert to_correct.attrs['reference_date'] == datetime(2019, 1, 1)
+        assert to_correct.short_wavelength_displacement.values == 13